ParamMethod Pane updates its dependencies dynamically
