+++ panel/pane.py
@@ -232,76 +232,6 @@ class HoloViews(PaneBase):
         return model
 
 
-class ParamMethod(PaneBase):
-    """
-    ParamMethod panes wrap methods annotated with the param.depends
-    decorator and rerenders the plot when any of the methods parameters
-    change. The method may return any object which itself can be rendered
-    as a Pane.
-    """
-
-    def __init__(self, object, **params):
-        self._kwargs =  {p: params.pop(p) for p in list(params)
-                         if p not in self.params()}
-        super(ParamMethod, self).__init__(object, **params)
-        self._pane = Pane(self.object(), name=self.name,
-                          **dict(_temporary=True, **self._kwargs))
-
-    @classmethod
-    def applies(cls, obj):
-        return inspect.ismethod(obj) and isinstance(get_method_owner(obj), param.Parameterized)
-
-    def _get_model(self, doc, root=None, parent=None, comm=None):
-        parameterized = get_method_owner(self.object)
-        params = parameterized.param.params_depended_on(self.object.__name__)
-        model = self._pane._get_model(doc, root, parent, comm)
-        ref = model.ref['id']
-        history = [model]
-        def update_pane(*change, history=history):
-            # Try updating existing pane
-            old_model = history[0]
-            new_object = self.object()
-            pane_type = self.get_pane_type(new_object)
-            if type(self._pane) is pane_type:
-                if isinstance(new_object, PaneBase):
-                    new_params = {k: v for k, v in new_object.get_param_values()
-                                  if k != 'name'}
-                    self._pane.set_param(**new_params)
-                    new_object._cleanup(None, new_object._temporary)
-                else:
-                    self._pane.object = new_object
-                return
-
-            # Replace pane entirely
-            self._pane._cleanup(old_model, self._pane._temporary)
-            self._pane = Pane(new_object, _temporary=True, **self._kwargs)
-            new_model = self._pane._get_model(doc, root, parent, comm)
-            def update_models():
-                if old_model is new_model: return
-                index = parent.children.index(old_model)
-                parent.children[index] = new_model
-                history[0] = new_model
-
-            if comm:
-                update_models()
-                push(doc, comm)
-            else:
-                doc.add_next_tick_callback(update_models)
-
-        for _, params in full_groupby(params, lambda x: (x.inst or x.cls, x.what)):
-            p = params[0]
-            pobj = (p.inst or p.cls)
-            ps = [p.name for p in params]
-            watcher = pobj.param.watch(update_pane, ps, p.what)
-            self._callbacks[ref].append(watcher)
-
-        return model
-
-    def _cleanup(self, model=None, final=False):
-        self._pane._cleanup(model, final)
-        super(ParamMethod, self)._cleanup(model, final)
-
-
 class DivPaneBase(PaneBase):
     """
     Baseclass for Panes which render HTML inside a Bokeh Div.
